+++ lib/lightning_web/live/run_live/streaming.ex
@@ -52,8 +52,11 @@ defmodule LightningWeb.RunLive.Streaming do
       nil ->
         {nil, []}
 
-      %Dataclip{id: id, body: body, type: type} ->
-        {%{id: id, step_id: step.id, type: type},
+      %Dataclip{body: nil} ->
+        {nil, []}
+
+      %Dataclip{id: id, body: body, type: type, wiped_at: wiped_at} ->
+        {%{id: id, step_id: step.id, type: type, wiped_at: wiped_at},
          body
          |> Jason.encode!(pretty: true)
          |> maybe_scrub(type, step)